Patch button focus when clicked on WebKit
Webkit doesn't give a `focus` state to buttons when clicked (a behaviour that is different from all the other vendors; see https://bugs.webkit.org/show_bug.cgi?id=22261). To overcome this difference in behaviour we apply a focus state to the buttons used to toggle the disclosed content.

Fixup! Patch button focus when clicked WebKit