Fix badge-description spacing in RadioCard
When a badge exists, the spacing between badge and description is `12px`.

When the badge doesn't exist the spacing between label and description is `8px`

To match these rules we adjust the CSS properties accordingly and try to rely on margin-bottom only, especially when dealing with spacing between optional elements.