Update ci-components-nightly.yml
Now that we know the cron job works (yay!) and that it produces the artifact that we want in the format that we want (double yay!), we have to next solve _how_ we're going to use the information returned from the test results.  

So while we figure that out, I'm turning off the cron part of the job; it can still be run manually.