Remove aria-hidden from optional span
Aria hidden was added to the isRequired badge because the input element itself is denoted as required and therefore having it on the badge too is repetitive for SRs and other AT. However, this is not true for the optional span. The aria-hidden attribute was likely added accidentally at the same time.